boards: Add support for active high buttons
The button on the PineTime isn't a simple pull low design. Instead
the switch is connected across two GPIO pins *and* has external
pull downs on both pins. This circuit requires special handling since
it must be active high and we need to configure the other pin as an
output for the pin to work correctly.