boards: Extend the watchdog to cover the bootloader too
If the bootloader fails for any reason we'd like to be able to restart
it! This is approach is a bit of a cheat... we are feeding the dog from
a periodic interrupt and there are plenty of ways to the system to fail
that leave periodic interrupts still ticking. However since we test the
user button and reboot from the ISR then this is OK... either the SysTick
stops and we reboot or the SysTick is running and the reboot switch will
still work.